+++ cmd/gocolor/main.go
@@ -9,10 +9,18 @@ import (
 
 func main() {
 	var (
-		cli = cmdline.NewBasicParser()
+		cli    = cmdline.NewBasicParser()
+		custom = cli.Option("-c, --custom").String("")
+	)
+	usage := cli.Usage()
+	usage.Example("custom color by regexp",
+		`$ gocolor -c "error.*:red info.*:green"`,
+		"",
+		"custom value is a space delimited list of REGEXP:COLOR",
+		"",
 	)
 	cli.Parse()
-	err := gocolor.Colorize(os.Stdout, os.Stdin)
+	err := gocolor.Colorize(os.Stdout, os.Stdin, custom)
 	if err != nil {
 		os.Exit(1)
 	}